requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.3.10
openai-whisper>=20231117
python-dotenv>=1.0.0
pyperclip>=1.8.2
requests>=2.31.0
